docstrings, try/raises standardization

+++ etl/geodata.py
@@ -28,19 +28,19 @@ def fetch_geodata(year: int) -> gpd.GeoDataFrame:
         Exception: If fetching geodata fails.
     """
     try:
-        gdf = geobr.read_municipality(code_muni = 'all',
+        gdf = geobr.read_municipality(code_muni = "all",
                                       year = year)
-        gdf = gpd.GeoDataFrame(gdf).drop(columns = ['name_muni',
-                                                    'code_state',
-                                                    'abbrev_state'])
+        gdf = gpd.GeoDataFrame(gdf).drop(columns = ["name_muni",
+                                                    "code_state",
+                                                    "abbrev_state"])
         return gdf
     except Exception as e:
-        logging.error(f'Error fetching geobr data: {e}')
+        logging.error(f"Error fetching geobr data: {e}")
         raise
 
 def merge_geodata(data : pd.DataFrame,
                   geodata : gpd.GeoDataFrame) -> gpd.GeoDataFrame:
-        """
+    """
     Merge the analytical dataset with municipality geodata.
     Args:
         data (pd.DataFrame): Silver-layer analytical dataset with 'city_id' column.
@@ -50,23 +50,23 @@ def merge_geodata(data : pd.DataFrame,
     Raises:
         Exception: If merging geodata fails.
     """
-        try:
-            df = data.copy()
-            df['city_id_int'] = df['city_id'].astype(str).astype(int)
-            geodata['code_muni'] = geodata['code_muni'].astype(int)
+    try:
+        df = data.copy()
+        df["city_id_int"] = df["city_id"].astype(str).astype(int)
+        geodata["code_muni"] = geodata["code_muni"].astype(int)
 
-            merged_map = df.merge(geodata[["code_muni", "geometry"]],
-                                  how="left",
-                                  left_on="city_id_int",
-                                  right_on="code_muni")
-            merged_map = merged_map.drop(columns=["city_id_int", "code_muni"])
+        merged_map = df.merge(geodata[["code_muni", "geometry"]],
+                              how="left",
+                              left_on="city_id_int",
+                              right_on="code_muni")
+        merged_map = merged_map.drop(columns=["city_id_int", "code_muni"])
 
-            merged_map = gpd.GeoDataFrame(merged_map,
-                                          geometry = 'geometry',
-                                          crs = geodata.crs)
-            merged_map['geometry'] = merged_map.geometry.simplify(tolerance = 0.01)
-            print("geobr data merging completed")
-            return merged_map
-        except Exception as e:
-            logging.error(f'Error merging geobr data: {e}')
-            raise
+        merged_map = gpd.GeoDataFrame(merged_map,
+                                      geometry = "geometry",
+                                      crs = geodata.crs)
+        merged_map["geometry"] = merged_map.geometry.simplify(tolerance = 0.01)
+        print("geobr data merging completed")
+        return merged_map
+    except Exception as e:
+        logging.error(f"Error merging geobr data: {e}")
+        raise
